chatbot/requirements.txt
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-groq
faiss-cpu
pypdf
python-dotenv
pydantic
fastapi
uvicorn